+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-- 2026-??-?? 2.3.21
+- 2026-04-09 2.3.21
     - XS Area Scan
         - fixed line increment
         - disable area scan while changing line increment